Bugfix in stubextractor

+++ src/StubExtractor.cc
@@ -41,7 +41,7 @@ StubExtractor::StubExtractor(bool doTree)
   m_stub_ladder  = new  std::vector<int>; 
   m_stub_seg     = new  std::vector<int>;  
   m_stub_chip    = new  std::vector<int>;  
-  m_stub_strip   = new  std::vector<int>; 
+  m_stub_strip   = new  std::vector<float>; 
   m_stub_x       = new  std::vector<float>;  
   m_stub_y       = new  std::vector<float>;  
   m_stub_z       = new  std::vector<float>;  
@@ -156,7 +156,7 @@ StubExtractor::StubExtractor(TFile *a_file)
   m_stub_module  = new  std::vector<int>;  
   m_stub_ladder  = new  std::vector<int>; 
   m_stub_seg     = new  std::vector<int>;  
-  m_stub_strip   = new  std::vector<int>; 
+  m_stub_strip   = new  std::vector<float>; 
   m_stub_chip    = new  std::vector<int>; 
   m_stub_x       = new  std::vector<float>;  
   m_stub_y       = new  std::vector<float>;
